#17 Orders http://pm.5psolutions.net/public/index.php?path_info=projects/training-teodora-acatrinei/tasks/17 replaced float columns in db with decimal

+++ database/migrations/2022_07_20_082348_create_order_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('order_id')->constrained();
             $table->foreignId('product_id')->constrained();
-            $table->float('product_price', 8, 2);
+            $table->decimal('product_price', 8, 2);
             $table->timestamps();
         });
     }
